partition actions
