+++ lib/src/redux/actions/actions_activity.dart
@@ -1,0 +1,16 @@
+class SubmitActivityAttempt {
+  String type;
+  String title;
+  String body;
+  String idCabildo;
+  String tags;
+  SubmitActivityAttempt(this.type, this.title, this.body, this.idCabildo, this.tags);
+}
+
+class SubmitActivitySuccess {
+}
+
+class SubmitActivityError {
+  String error;
+  SubmitActivityError(this.error);
+}
