+++ lib/src/redux/actions/actions_user.dart
@@ -1,0 +1,25 @@
+import 'package:cibic_mobile/src/models/feed_model.dart';
+import 'package:cibic_mobile/src/models/user_model.dart';
+
+class FetchUserProfileAttempt {
+}
+
+class FetchUserProfileSuccess {
+  UserModel user;
+  FetchUserProfileSuccess(this.user);
+}
+
+class FetchUserProfileError {
+  String error;
+  FetchUserProfileError(this.error);
+}
+
+class FetchUserProfileFeedSuccess {
+  FeedModel feed;
+  FetchUserProfileFeedSuccess(this.feed);
+}
+
+class FetchUserProfileFeedError {
+  String error;
+  FetchUserProfileFeedError(this.error);
+}
