+++ lib/src/redux/middleware/thunk.dart
@@ -1,3 +1,6 @@
+import 'package:cibic_mobile/src/redux/actions/actions_activity.dart';
+import 'package:cibic_mobile/src/redux/actions/actions_cabildo.dart';
+import 'package:cibic_mobile/src/redux/actions/actions_user.dart';
 import 'package:cibic_mobile/src/redux/middleware/thunk_compose.dart';
 import 'package:cibic_mobile/src/redux/middleware/thunk_feed.dart';
 import 'package:cibic_mobile/src/redux/middleware/thunk_menu_overlay.dart';
